requirements.txt
requests
pillow
youtube_dl